wip - move notebook integration test into our normal API test extension
